+++ test/run_all_versions.sh
@@ -61,6 +61,7 @@ SUITES=(
 	bloom_setting
 	bloom_sizing
 	cancel_decode
+	catalog_delete_index
 	catalog_natts
 	catalog_plan_index
 	column_projection